feat(db): trades price/total_amount/profit_loss_amount Numeric(18,4) 전환 (P3c-3, 운영자 마이그 필요)
멀티마켓 정밀 저장 — KRX 정수 무손실(.0000), US 센트 보존. asdecimal=False로
Python 읽기는 float 유지(기존 집계·JSON 직렬화 호환, Decimal 타입혼합 회피).

- src/db/models.py: Trade.price/total_amount/profit_loss_amount Integer→Numeric(18,4).
- alembic c3d4e5f6a7b8: Integer→numeric(18,4) USING 캐스트(비파괴), down=round()::integer.
  → 운영자 액션: alembic upgrade head (P3b 마이그가 예고한 후속).
- src/db/repository.record_trade: 시그니처 int→float. DailySummary(정수 KRX 컬럼)는 int 캐스트(비범위).
- src/db/analytics.py: PnL 누적 dict/list 타입 int→float(Numeric 읽기 정합).
- src/engine.py: _record_trade_to_db profit_loss_amount를 _norm_price로(US 센트 DB 보존).
  RiskManager.record_trade_result 입력은 int 유지(연패/MDD 게이트 KRX 불변).
- tests: Numeric 컬럼 타입 + US 센트 round-trip + KRX 정수 무손실 검증(신규 3).

검증: pytest 1163 passed(신규 3) | mypy strict 102 files | ruff 변경분 clean.

+++ src/db/analytics.py
@@ -731,12 +731,12 @@ def get_cumulative_pnl(
         .order_by(Trade.traded_at)
     ).scalars().all()
 
-    daily: dict[str, int] = {}
+    daily: dict[str, float] = {}
     for t in sells:
         day_key = t.traded_at.strftime("%Y-%m-%d")
         daily[day_key] = daily.get(day_key, 0) + (t.profit_loss_amount or 0)
 
-    cumulative = 0
+    cumulative = 0.0
     curve = []
     for day_key in sorted(daily):
         cumulative += daily[day_key]
@@ -804,7 +804,7 @@ def get_strategy_comparison(
         )
     ).scalars().all()
 
-    pnl_by_signal: dict[str, list[int]] = {}
+    pnl_by_signal: dict[str, list[float]] = {}
     for t in sells:
         sig = t.signal_type or "UNKNOWN"
         if sig not in pnl_by_signal:
@@ -912,7 +912,7 @@ def get_optimal_risk_params(
 
 def _daily_pnl_series(
     session: Session, start_date: date, end_date: date,
-) -> list[tuple[str, int]]:
+) -> list[tuple[str, float]]:
     """기간 내 일별 실현손익 시계열을 반환한다 (내부 헬퍼)."""
     start = datetime(start_date.year, start_date.month, start_date.day)
     end = datetime(end_date.year, end_date.month, end_date.day) + timedelta(days=1)
@@ -927,7 +927,7 @@ def _daily_pnl_series(
         .order_by(Trade.traded_at)
     ).scalars().all()
 
-    daily: dict[str, int] = {}
+    daily: dict[str, float] = {}
     for t in sells:
         day_key = t.traded_at.strftime("%Y-%m-%d")
         daily[day_key] = daily.get(day_key, 0) + (t.profit_loss_amount or 0)
@@ -954,12 +954,12 @@ def get_max_drawdown(
     if not series:
         return {"mdd_pct": 0.0, "mdd_amount": 0, "peak": 0, "trough": 0, "peak_date": None, "trough_date": None}
 
-    cumulative = 0
-    peak = 0
+    cumulative = 0.0
+    peak = 0.0
     peak_date = series[0][0]
-    mdd_amount = 0
-    mdd_peak = 0
-    mdd_trough = 0
+    mdd_amount = 0.0
+    mdd_peak = 0.0
+    mdd_trough = 0.0
     mdd_peak_date = series[0][0]
     mdd_trough_date = series[0][0]
